+++ sql/core/src/test/scala/org/apache/spark/sql/execution/streaming/state/StateStoreSuite.scala
@@ -36,16 +36,12 @@ import org.scalatest.time.SpanSugar._
 import org.apache.spark._
 import org.apache.spark.LocalSparkContext._
 import org.apache.spark.internal.config.Network.RPC_NUM_RETRIES
-import org.apache.spark.io.CompressionCodec
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.catalyst.expressions.{GenericInternalRow, UnsafeProjection, UnsafeRow}
-import org.apache.spark.sql.catalyst.plans.PlanTestBase
-import org.apache.spark.sql.catalyst.streaming.InternalOutputModes.Update
 import org.apache.spark.sql.catalyst.util.quietly
 import org.apache.spark.sql.execution.streaming._
 import org.apache.spark.sql.functions.count
 import org.apache.spark.sql.internal.SQLConf
-import org.apache.spark.sql.streaming.StreamTest
 import org.apache.spark.sql.types._
 import org.apache.spark.unsafe.types.UTF8String
 import org.apache.spark.util.Utils
@@ -815,48 +811,6 @@ class StateStoreSuite extends StateStoreSuiteBase[HDFSBackedStateStoreProvider]
     val filePath = new File(basePath.toString, fileName)
     filePath.delete()
     filePath.createNewFile()
-  }
-}
-
-class StateStoreCompatibleSuite extends StreamTest with StateStoreCodecsTest {
-  testWithAllCodec("SPARK-33263: Recovery from checkpoint before codec config introduced") {
-    val resourceUri = this.getClass.getResource(
-      "/structured-streaming/checkpoint-version-3.0.0-streaming-statestore-codec/").toURI
-    val checkpointDir = Utils.createTempDir().getCanonicalFile
-    FileUtils.copyDirectory(new File(resourceUri), checkpointDir)
-
-    import testImplicits._
-
-    val inputData = MemoryStream[Int]
-    val aggregated = inputData.toDF().groupBy("value").agg(count("*"))
-    inputData.addData(1, 2, 3)
-
-    /*
-      Note: The checkpoint was generated using the following input in Spark version 3.0.0:
-      AddData(inputData, 1, 2, 3)
-    */
-    testStream(aggregated, Update)(
-      StartStream(
-        checkpointLocation = checkpointDir.getAbsolutePath,
-        additionalConfs = Map(SQLConf.SHUFFLE_PARTITIONS.key -> "1")),
-      AddData(inputData, 1, 2),
-      CheckNewAnswer((1, 2), (2, 2))
-    )
-  }
-}
-
-trait StateStoreCodecsTest extends SparkFunSuite with PlanTestBase {
-  private val allCodecs =
-    CompressionCodec.ALL_COMPRESSION_CODECS.map { c => CompressionCodec.getShortName(c) }
-
-  protected def testWithAllCodec(name: String)(func: => Any): Unit = {
-    allCodecs.foreach { codecShortName =>
-      withSQLConf(SQLConf.STATE_STORE_COMPRESSION_CODEC.key -> codecShortName) {
-        test(s"$name - with codec $codecShortName") {
-          func
-        }
-      }
-    }
   }
 }
